feat(storage): move common client methods to abstract client class
